Task5/requirements.txt
pymongo>=4.5.0
python-dotenv>=1.0.0
matplotlib==3.9.0
psutil==5.9.8
